+++ main.go
@@ -155,6 +155,7 @@ func (br *WABridge) Start() {
 		br.Provisioning.Init()
 	}
 	go br.CheckWhatsAppUpdate()
+	go br.UpdatePuppetContactInfo()
 	go br.StartUsers()
 	br.UpdateActivePuppetCount()
 	if br.Config.Metrics.Enabled {
@@ -183,6 +184,14 @@ func (br *WABridge) CheckWhatsAppUpdate() {
 		}
 	} else {
 		br.Log.Debugfln("Bridge is using newer than latest WhatsApp web protocol")
+	}
+}
+
+func (br *WABridge) UpdatePuppetContactInfo() {
+	for _, puppet := range br.GetAllPuppets() {
+		if puppet.UpdateContactInfo() {
+			puppet.Update()
+		}
 	}
 }
 
@@ -282,11 +291,13 @@ func main() {
 		},
 	}
 	br.Bridge = bridge.Bridge{
-		Name:         "mautrix-whatsapp",
-		URL:          "https://github.com/mautrix/whatsapp",
-		Description:  "A Matrix-WhatsApp puppeting bridge.",
-		Version:      "0.8.3-mod-1",
-		ProtocolName: "WhatsApp",
+		Name:              "mautrix-whatsapp",
+		URL:               "https://github.com/mautrix/whatsapp",
+		Description:       "A Matrix-WhatsApp puppeting bridge.",
+		Version:           "0.8.4-mod-1",
+		ProtocolName:      "WhatsApp",
+		BeeperServiceName: "whatsapp",
+		BeeperNetworkName: "whatsapp",
 
 		CryptoPickleKey: "maunium.net/go/mautrix-whatsapp",