+++ puppet.go
@@ -28,6 +28,7 @@ import (
 
 	"maunium.net/go/mautrix/appservice"
 	"maunium.net/go/mautrix/bridge"
+	"maunium.net/go/mautrix/bridge/bridgeconfig"
 	"maunium.net/go/mautrix/id"
 
 	"maunium.net/go/mautrix-whatsapp/config"
@@ -267,30 +268,57 @@ func (puppet *Puppet) UpdateName(contact types.ContactInfo, forcePortalSync bool
 	return false
 }
 
+func (puppet *Puppet) UpdateContactInfo() bool {
+	if puppet.bridge.Config.Homeserver.Software != bridgeconfig.SoftwareHungry {
+		return false
+	}
+
+	if puppet.ContactInfoSet {
+		return false
+	}
+
+	contactInfo := map[string]any{
+		"com.beeper.bridge.identifiers": []string{
+			fmt.Sprintf("tel:+%s", puppet.JID.User),
+			fmt.Sprintf("whatsapp:%s", puppet.JID.String()),
+		},
+		"com.beeper.bridge.remote_id":     puppet.JID.String(),
+		"com.beeper.bridge.service":       "whatsapp",
+		"com.beeper.bridge.network":       "whatsapp",
+		"com.beeper.bridge.is_bridge_bot": false,
+		"com.beeper.bridge.is_bot":        false,
+	}
+	err := puppet.DefaultIntent().BeeperUpdateProfile(contactInfo)
+	if err != nil {
+		puppet.log.Warnln("Failed to store custom contact info in profile:", err)
+		return false
+	} else {
+		puppet.ContactInfoSet = true
+		return true
+	}
+}
+
 func (puppet *Puppet) updatePortalMeta(meta func(portal *Portal)) {
-	if puppet.bridge.Config.Bridge.PrivateChatPortalMeta || puppet.bridge.Config.Bridge.Encryption.Allow {
-		for _, portal := range puppet.bridge.GetAllPortalsByJID(puppet.JID) {
-			if !puppet.bridge.Config.Bridge.PrivateChatPortalMeta && !portal.Encrypted {
-				continue
-			}
-			// Get room create lock to prevent races between receiving contact info and room creation.
-			portal.roomCreateLock.Lock()
-			meta(portal)
-			portal.roomCreateLock.Unlock()
-		}
+	for _, portal := range puppet.bridge.GetAllPortalsByJID(puppet.JID) {
+		// Get room create lock to prevent races between receiving contact info and room creation.
+		portal.roomCreateLock.Lock()
+		meta(portal)
+		portal.roomCreateLock.Unlock()
 	}
 }
 
 func (puppet *Puppet) updatePortalAvatar() {
 	puppet.updatePortalMeta(func(portal *Portal) {
-		if portal.Avatar == puppet.Avatar && portal.AvatarURL == puppet.AvatarURL && portal.AvatarSet {
+		if portal.Avatar == puppet.Avatar && portal.AvatarURL == puppet.AvatarURL && (portal.AvatarSet || !portal.shouldSetDMRoomMetadata()) {
 			return
 		}
 		portal.AvatarURL = puppet.AvatarURL
 		portal.Avatar = puppet.Avatar
 		portal.AvatarSet = false
 		defer portal.Update(nil)
-		if len(portal.MXID) > 0 {
+		if len(portal.MXID) > 0 && !portal.shouldSetDMRoomMetadata() {
+			portal.UpdateBridgeInfo()
+		} else if len(portal.MXID) > 0 {
 			_, err := portal.MainIntent().SetRoomAvatar(portal.MXID, puppet.AvatarURL)
 			if err != nil {
 				portal.log.Warnln("Failed to set avatar:", err)
@@ -343,6 +371,7 @@ func (puppet *Puppet) Sync(source *User, contact *types.ContactInfo, forceAvatar
 	if len(puppet.Avatar) == 0 || forceAvatarSync || puppet.bridge.Config.Bridge.UserAvatarSync {
 		update = puppet.UpdateAvatar(source, forcePortalSync) || update
 	}
+	update = puppet.UpdateContactInfo() || update
 	if update || puppet.LastSync.Add(24*time.Hour).Before(time.Now()) {
 		puppet.LastSync = time.Now()
 		puppet.Update()